# api-docs/routes/openapi.yaml
openapi: 3.0.1
info:
  title: UZH Masters Project API Documentation
  description: "The following list of API endpoints connected by the API Gateway to the different Microservices. Click [Try it out] to test the endpoints and edit the Example Value."
  contact:
    email: calvinmitchell.ongkingco@uzh.ch
  version: 1.14.3

# Gateway Link
servers:
- url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/
  description: The gateway link of all endpoint services

# Tags for the endpoint services.
...
paths:
...
  /marker:
    post:
      tags:
      - mapping
      summary: Create a map marker
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      description: Creates a map marker
      operationId: createMapMarker
      requestBody:
        description: Sample details to create a map marker. Click "Try it out" to edit the values
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMapMarker'
        required: true
      responses:
        "200":
          description: Returns the JSON Created.
        "500":
          description: "[ERROR] Error in Inserting Mapping Details to Database. error"
      x-swagger-router-controller: Mapping
      security:
      - tokenAuth: []
    put:
      tags:
      - mapping
      summary: Update map marker. [WEB APP]
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      description: Updates a map marker
      operationId: updateMapMarker
      requestBody:
        description: Update a map marker
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMapMarker'
        required: true
      responses:
        "200":
          description: Successfully Updated Marker Details.
        "404":
          description: No Record Found Matching Map ID.
        "500":
          description: Error saving record
      security:
      - tokenAuth: []
      x-swagger-router-controller: Mapping
  /marker/{id}:
    get:
      tags:
      - mapping
      summary: Get details of map marker
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      description: Get details of map marker
      operationId: getMapMarkerDetails
      parameters:
      - name: id
        in: path
        description: ID of map marker to get details
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "200":
          description: "[SUCCESS] Sucessfully Returning Map Details of (marker)"
        "404":
          description: No Marker Record Found
        "500":
          description: Error in Initializing Connection to Database.
      x-swagger-router-controller: Mapping
      security:
      - tokenAuth: []
    delete:
      tags:
      - mapping
      summary: Delete map marker [WEB APP]
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      description: Deletes a map marker
      operationId: removeMapMarker
      parameters:
      - name: id
        in: path
        description: ID of map marker to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "200":
          description: Successfully Deleted Marker
        "404":
          description: No Marker Record Found
        "500":
          description: Error in Initializing Connection to Database
      x-swagger-router-controller: Mapping
      security:
      - tokenAuth: []
  /marker/quest:
    put:
      tags:
      - mapping
      summary: Verify Quest Markers
      description: Updates the Existing Quest Marker Record
      operationId: verifyingQuests
      requestBody:
        description: Sample Details of Verifying a Quest Marker, Rename the mid to markerid. Click "Try it out" to edit the values
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyingQuests'
        required: true
      parameters:
      - name: id
        in: path
        description: MID of map marker to Verify Quest
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      responses:
        "200":
          description: Successfully Created New Quest
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Mapping
      security:
      - tokenAuth: []
  /open/markers:
    get:
      tags:
      - mapping
      summary: Get All Unverified Quests
      description: A function to request all open or unverified quests.
      operationId: getUnverifiedQuests
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg/msrv
      responses:
        "200":
          description: Returns Unverified Quest Marker JSON
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Mapping
      security:
      - tokenAuth: []

 
# -------------------------------------------------------------
components:
  schemas:
...
    CreateMapMarker:
      type: object
      additionalProperties: false
      properties:
        geo_type:
          type: string
          description: Location type marker
          enum:
          - point
          - polygon
          - node
          - way
          - relation
        lat:
          type: integer
          description: Latitude
        long:
          type: integer
          description: Longitude
        nodeid:
          type: string
          description: OSM Node ID
        version:
          type: integer
          description: Current version of the node
        title:
          type: string
          description: Title of Quest
        subtitle:
          type: string
          description: Subtitle of the Quest
        description:
          type: string
          description: Description of the Quest 
        comments:
          type: string
          description: Comment to be passed to OSM
        tags:
          type: array
          items:
            type: string
            properties:
              id:
                type: integer
              name:
                type: string
          example: [{ "k":"Tag Name","v":"Tag Value" },{ "k":"Tag Name","v":"Tag Value" }]
        image_url:
          type: array
          items:
            type: string
            properties:
              image_url:
                type: string
          example: ["some url","another url"]
    UpdateMapMarker:
      type: object
      additionalProperties: false
      properties:
        markerid:
          type: string
          description: The map marker's ID
        loc_type:
          type: string
          description: Location type marker
          enum:
          - point
          - polygon
        lat:
          type: integer
          description: Latitude
        long:
          type: integer
          description: Longitude
        title:
          type: string
          description: Please Add a valid Location Title
        subtitle:
          type: string
          description: Please Add a valid subtitle 
        description:
          type: string
          description: Please Add a description
    VerifyingQuests:
      type: object
      properties:
        markerid:
          type: string
          description: Verify Quest
        description:
          type: string
          description: Add User Comment
        image_url:
          type: string
          description: Image URL        
      additionalProperties: false
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT